Make nox easier to interrupt + move files around.
