+++ tests/performance/assignment_chain_performance.lp
