+++ tests/performance/sum_aggregates.lp
